# AspNetCoreDiagnosticObserver optimizations.