- Adds DuckTyping mechanism.
- Adds ReferenceEquals optimization

#### Benchmark for the ReferenceEquals: (*Changed*)

```ini
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.630 (2004/?/20H1)
Intel Core i7-1068NG7 CPU 2.30GHz, 1 CPU, 2 logical and 2 physical cores
.NET Core SDK=5.0.100
  [Host]     : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT
  Job-AIVVBG : .NET Framework 4.8 (4.8.4250.0), X64 RyuJIT
  Job-KHFMMM : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT
```

|           Method |        Job |       Runtime |     Toolchain |     Mean |   Error |  StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |----------- |-------------- |-------------- |---------:|--------:|--------:|------:|-------:|------:|------:|----------:|
|    CurrentOnNext | Job-AIVVBG |    .NET 4.7.2 |        net472 | 185.0 ns | 1.43 ns | 1.19 ns |  1.00 | 0.0668 |     - |     - |     281 B |
|    CurrentOnNext | Job-KHFMMM | .NET Core 3.1 | netcoreapp3.1 | 183.4 ns | 0.76 ns | 0.63 ns |  0.99 | 0.0668 |     - |     - |     280 B |
|                  |            |               |               |          |         |         |       |        |       |       |           |
| CurrentOnNextNew | Job-AIVVBG |    .NET 4.7.2 |        net472 | 178.4 ns | 0.92 ns | 0.86 ns |  1.00 | 0.0668 |     - |     - |     281 B |
| CurrentOnNextNew | Job-KHFMMM | .NET Core 3.1 | netcoreapp3.1 | 174.6 ns | 0.94 ns | 0.88 ns |  0.98 | 0.0668 |     - |     - |     280 B |


@DataDog/apm-dotnet